fixed a couple gcc/linux-specific changes for new strong enum support (stuff that was ifdef posix)

+++ Library/Sources/Stroika/Foundation/Execution/Logger.cpp
@@ -83,7 +83,7 @@ Logger::SysLogAppender::~SysLogAppender ()
 void    Logger::SysLogAppender::Log (Priority logLevel, const String& message) override
 {
     DbgTrace (L"%s", message.c_str ());
-    syslog (logLevel, "%s", message.AsTString ().c_str ());
+    syslog (static_cast<int> (logLevel), "%s", message.AsTString ().c_str ());
 }
 #endif
